+++ src/renderer/src/constants/env.ts
@@ -1,0 +1,2 @@
+export const isDevMode = import.meta.env.DEV && import.meta.env.MODE === "development"
+export const isProdMode = import.meta.env.PROD && import.meta.env.MODE === "production"