+++ src/renderer/src/ui/modules/Settings/Settings.vue
@@ -1,13 +1,14 @@
 <script setup lang="ts">
+import {isDevMode} from "@/constants/env"
 import BaseIcon from "@/ui/base/BaseIcon"
 import ExpansionPanel from "@/ui/common/misc/ExpansionPanel/ExpansionPanel.vue"
 
 import About from "./{fragments}/About.vue"
 import AiSettings from "./{fragments}/AiSettings"
 import DeletedTasks from "./{fragments}/DeletedTasks"
+import IconsList from "./{fragments}/IconsList.vue"
 import SyncSettings from "./{fragments}/SyncSettings.vue"
 import Themes from "./{fragments}/Themes.vue"
-import IconsList from "../IconsList"
 
 const COMMON_GROUP = "settings"
 </script>
@@ -57,14 +58,16 @@ const COMMON_GROUP = "settings"
       <DeletedTasks />
     </ExpansionPanel>
 
-    <ExpansionPanel :group="COMMON_GROUP">
-      <template #header>
-        <h3 class="text-base-content flex items-center gap-2 text-sm font-semibold">
-          <BaseIcon name="heading" class="text-accent size-4" />
-          Icons
-        </h3>
-      </template>
-      <IconsList />
-    </ExpansionPanel>
+    <template v-if="isDevMode">
+      <ExpansionPanel :group="COMMON_GROUP">
+        <template #header>
+          <h3 class="text-base-content flex items-center gap-2 text-sm font-semibold">
+            <BaseIcon name="heading" class="text-accent size-4" />
+            Icons
+          </h3>
+        </template>
+        <IconsList />
+      </ExpansionPanel>
+    </template>
   </div>
 </template>